+++ packages/ui/decorators/index.ts
@@ -1,2 +1,3 @@
 export * from './withTransition.js';
 export * from './withDeprecation.js';
+export * from './withIndex.js';